Warn "unsupported software" on start
While the README file notes that grace-gtk is totally unsupported, users
may miss that passage and be misled about the nature of the software.
With this commit a warning message will be displayed when the library is
used to ensure the user knows this and to point them at the source code
repository in case they want to contribute improvements.